new terminal manual
saved from a lost copy of flossmanual's command-line manual, which was also an
FSF book, but is not available anymore online, only from archive.org

downloaded from html along with images, sanitized, converted to markdown and
polished a bit (was a mess with repeating chapters). Still some polishing to be
done about escaped chars, but its at a good point now.